streamingccl: minor refactor
This pulls out the per-row handling into an interface that we can test
separately.  The goal here is to cut down a bit on conflicts as we
iterate on the batching strategy separately from the underlying insert
queries.

Epic: none
Release note: none
