+++ pkg/ccl/streamingccl/logical/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
         "logical_replication_dist.go",
         "logical_replication_job.go",
         "logical_replication_writer_processor.go",
+        "lww_row_processor.go",
         "metrics.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/ccl/streamingccl/logical",
@@ -20,6 +21,7 @@ go_library(
         "//pkg/jobs",
         "//pkg/jobs/jobspb",
         "//pkg/jobs/jobsprofiler",
+        "//pkg/keys",
         "//pkg/repstream/streampb",
         "//pkg/roachpb",
         "//pkg/settings",
@@ -27,6 +29,7 @@ go_library(
         "//pkg/sql",
         "//pkg/sql/catalog",
         "//pkg/sql/catalog/descpb",
+        "//pkg/sql/catalog/descs",
         "//pkg/sql/catalog/tabledesc",
         "//pkg/sql/execinfra",
         "//pkg/sql/execinfrapb",
@@ -37,8 +40,6 @@ go_library(
         "//pkg/sql/sem/catid",
         "//pkg/sql/sem/eval",
         "//pkg/sql/sem/tree",
-        "//pkg/sql/sessiondata",
-        "//pkg/sql/sessiondatapb",
         "//pkg/sql/types",
         "//pkg/util/ctxgroup",
         "//pkg/util/hlc",
